requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiomysql==0.2.0
aiosignal==1.4.0
asyncio==4.0.0
asyncssh==2.21.1
attrs==25.3.0
boto3==1.40.75
botocore==1.40.75
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Deprecated==1.2.18
distro==1.9.0
frozenlist==1.7.0
hiredis==3.3.0
idna==3.10
iniconfig==2.3.0
jmespath==1.0.1
multidict==6.6.4
netaddr==1.3.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.1.0
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydevd-pycharm==252.26830.99
Pygments==2.19.2
PyMySQL==1.1.2
pyTelegramBotAPI==4.29.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qrcode==8.2
redis==7.1.0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==6.0.0
wrapt==1.17.3
yarl==1.20.1
